rename package for matchers update example with matchers make maven/junit work by providing junit-bom
